Move the one mistake and the new class into the BloomTest namespace
